Feature JediTerm
Implemented some example code for using JediTerm manually.  Includes some buttons for copying everything, clearing the terminal, and manipulating the background and foreground colors.  JediTerm comes with a tty specifically to hook up to Process objects instead of what I'm doing here but this is a fantastic sign for getting this working.